delete lexer package
