+++ 10/jack_analyzer/jack_analyzer.go
@@ -1,9 +1,9 @@
 package jack_analyzer
 
-import (
-	"fmt"
-)
-
-func Analyzer() {
-	fmt.Println("hgoehgoe")
+// Lexer has member necessary for parsing
+type Lexer struct {
+	input        string // code input
+	position     int
+	readPosition int
+	ch           byte
 }
