--- 10/lexer/lexer.go
@@ -1,9 +1,0 @@
-package lexer
-
-// Lexer has member necessary for parsing
-type Lexer struct {
-	input        string // code input
-	position     int
-	readPosition int
-	ch           byte
-}
